+++ terraform/hydrocron-dynamo.tf
@@ -22,6 +22,9 @@ resource "aws_dynamodb_table" "hydrocron-swot-reach-table" {
     projection_type    = "INCLUDE"
     non_key_attributes = ["reach_id", "collection_shortname", "collection_version", "crid", "cycle_id", "pass_id", "continent_id", "ingest_time"]
   }
+  point_in_time_recovery {
+    enabled = var.stage == "ops" ? true : false
+  }
 }
 
 resource "aws_dynamodb_table" "hydrocron-swot-node-table" {
@@ -48,5 +51,65 @@ resource "aws_dynamodb_table" "hydrocron-swot-node-table" {
     range_key          = "range_start_time"
     projection_type    = "INCLUDE"
     non_key_attributes = ["node_id", "collection_shortname", "collection_version", "crid", "cycle_id", "pass_id", "continent_id", "ingest_time"]
+  }
+  point_in_time_recovery {
+    enabled = var.stage == "ops" ? true : false
+  }
+}
+
+resource "aws_dynamodb_table" "hydrocron-swot-prior-lake-table" {
+  name         = "hydrocron-swot-prior-lake-table"
+  billing_mode = "PAY_PER_REQUEST"
+  hash_key     = "lake_id"
+  range_key    = "range_start_time"
+  attribute {
+    name = "lake_id"
+    type = "S"
+  }
+  attribute {
+    name = "range_start_time"
+    type = "S"
+  }
+  attribute {
+    name = "granuleUR"
+    type = "S"
+  }
+
+  global_secondary_index {
+    name               = "GranuleURIndex"
+    hash_key           = "granuleUR"
+    range_key          = "range_start_time"
+    projection_type    = "INCLUDE"
+    non_key_attributes = ["lake_id", "collection_shortname", "collection_version", "crid", "cycle_id", "pass_id", "continent_id", "ingest_time"]
+  }
+  point_in_time_recovery {
+    enabled = var.stage == "ops" ? true : false
+  }
+}
+
+resource "aws_dynamodb_table" "hydrocron-track-ingest-table" {
+  name         = "hydrocron-track-ingest-table"
+  billing_mode = "PAY_PER_REQUEST"
+  hash_key     = "granuleUR"
+  range_key    = "revision_date"
+  attribute {
+    name = "granuleUR"
+    type = "S"
+  }
+  attribute {
+    name = "revision_date"
+    type = "S"
+  }
+  attribute {
+    name = "status"
+    type = "S"
+  }
+  global_secondary_index {
+    name               = "statusIndex"
+    hash_key           = "status"
+    projection_type    = "ALL"
+  }
+  point_in_time_recovery {
+    enabled = var.stage == "ops" ? true : false
   }
 }