+++ tests/test_hydrocron_database.py
@@ -11,7 +11,7 @@ See https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/DynamoDBLoc
 """
 from hydrocron.utils import constants
 
-from hydrocron.db.io import swot_reach_node_shp
+from hydrocron.db.io import swot_shp
 
 
 def test_table_exists(hydrocron_dynamo_table):
@@ -21,12 +21,12 @@ def test_table_exists(hydrocron_dynamo_table):
     assert hydrocron_dynamo_table.exists(constants.DB_TEST_TABLE_NAME)
 
 
-def test_add_data(hydrocron_dynamo_table):
+def test_add_data_reaches(hydrocron_dynamo_table):
     """
     Test adding data from one Reach shapefile to db
     """
-    items = swot_reach_node_shp.read_shapefile(
-        constants.TEST_SHAPEFILE_PATH,
+    items = swot_shp.read_shapefile(
+        constants.TEST_REACH_SHAPEFILE_PATH,
         obscure_data=False,
         columns=constants.REACH_DATA_COLUMNS)
 
@@ -40,8 +40,8 @@ def test_query(hydrocron_dynamo_table):
     """
     Test a query for a reach id
     """
-    items = swot_reach_node_shp.read_shapefile(
-        constants.TEST_SHAPEFILE_PATH,
+    items = swot_shp.read_shapefile(
+        constants.TEST_REACH_SHAPEFILE_PATH,
         obscure_data=False,
         columns=constants.REACH_DATA_COLUMNS)
 
@@ -51,17 +51,17 @@ def test_query(hydrocron_dynamo_table):
     items = hydrocron_dynamo_table.run_query(
         partition_key=constants.TEST_REACH_ID_VALUE)
 
-    assert items[0][constants.FIELDNAME_WSE] == constants.TEST_WSE_VALUE
-    assert items[0][constants.FIELDNAME_SWORD_VERSION] == constants.TEST_SWORD_VERSION_VALUE
-    assert items[0][constants.TEST_UNITS_FIELD] == constants.TEST_UNITS
+    assert items[0][constants.FIELDNAME_WSE] == constants.TEST_REACH_WSE_VALUE
+    assert items[0][constants.FIELDNAME_SWORD_VERSION] == constants.TEST_REACH_SWORD_VERSION_VALUE
+    assert items[0][constants.TEST_REACH_UNITS_FIELD] == constants.TEST_REACH_UNITS
 
 
 def test_delete_item(hydrocron_dynamo_table):
     """
     Test delete an item
     """
-    items = swot_reach_node_shp.read_shapefile(
-        constants.TEST_SHAPEFILE_PATH,
+    items = swot_shp.read_shapefile(
+        constants.TEST_REACH_SHAPEFILE_PATH,
         obscure_data=False,
         columns=constants.REACH_DATA_COLUMNS)
 
@@ -70,5 +70,5 @@ def test_delete_item(hydrocron_dynamo_table):
 
     hydrocron_dynamo_table.delete_item(
         partition_key=constants.TEST_REACH_ID_VALUE,
-        sort_key=constants.TEST_TIME_VALUE)
+        sort_key=constants.TEST_REACH_TIME_VALUE)
     assert hydrocron_dynamo_table.table.item_count == 686