+++ tests/test_io_swot_reach_node_shp.py
@@ -10,15 +10,15 @@ from datetime import datetime, timedelta, timezone
 import pytz
 from hydrocron.utils import constants
 
-from hydrocron.db.io import swot_reach_node_shp
+from hydrocron.db.io import swot_shp
 
 
-def test_parse_from_filename():
+def test_parse_from_filename_reach():
     """
     Tests parsing cycle, pass, and time ranges from filename
     """
-    filename_attrs = swot_reach_node_shp.parse_from_filename(
-        constants.TEST_FILENAME)
+    filename_attrs = swot_shp.parse_from_filename(
+        constants.TEST_REACH_FILENAME)
 
     assert filename_attrs['cycle_id'] == "548"
     assert filename_attrs['pass_id'] == "011"
@@ -28,34 +28,76 @@ def test_parse_from_filename():
     assert filename_attrs['crid'] == "PIA1"
     assert filename_attrs['collection_shortname'] == constants.SWOT_REACH_COLLECTION_NAME
     assert filename_attrs['collection_version'] == constants.SWOT_REACH_COLLECTION_VERSION
-    assert filename_attrs['granuleUR'] == constants.TEST_FILENAME
+    assert filename_attrs['granuleUR'] == constants.TEST_REACH_FILENAME
     assert datetime.strptime(filename_attrs['ingest_time'], "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=pytz.utc) - datetime.now(timezone.utc) <= timedelta(minutes=5)
 
 
-def test_read_shapefile():
+def test_parse_from_filename_lake():
+    """
+    Tests parsing cycle, pass, and time ranges from filename
+    """
+    filename_attrs = swot_shp.parse_from_filename(
+        constants.TEST_PLAKE_FILENAME)
+
+    assert filename_attrs['cycle_id'] == "018"
+    assert filename_attrs['pass_id'] == "100"
+    assert filename_attrs['continent_id'] == "GR"
+    assert filename_attrs['range_start_time'] == "2024-07-13T11:17:41Z"
+    assert filename_attrs['range_end_time'] == "2024-07-13T11:20:27Z"
+    assert filename_attrs['crid'] == "PIC0"
+    assert filename_attrs['collection_shortname'] == constants.SWOT_PRIOR_LAKE_COLLECTION_NAME
+    assert filename_attrs['collection_version'] == constants.SWOT_PRIOR_LAKE_COLLECTION_VERSION
+    assert filename_attrs['granuleUR'] == constants.TEST_PLAKE_FILENAME
+    assert datetime.strptime(filename_attrs['ingest_time'], "%Y-%m-%dT%H:%M:%SZ").replace(tzinfo=pytz.utc) - datetime.now(timezone.utc) <= timedelta(minutes=5)
+
+
+def test_read_reach_shapefile():
     """
     Tests reading attributes from the shapefile
     """
-    items = swot_reach_node_shp.read_shapefile(
-        constants.TEST_SHAPEFILE_PATH,
+    items = swot_shp.read_shapefile(
+        constants.TEST_REACH_SHAPEFILE_PATH,
         obscure_data=False,
         columns=constants.REACH_DATA_COLUMNS)
 
     assert len(items) == 687
-    for key, val in constants.TEST_ITEM_DICT.items():
+    for key, val in constants.TEST_REACH_ITEM_DICT.items():
         assert val == items[2][key]
+
+
+def test_read_lake_shapefile():
+    """
+    Tests reading attributes from the shapefile
+    """
+    items = swot_shp.read_shapefile(
+        constants.TEST_PLAKE_SHAPEFILE_PATH,
+        obscure_data=False,
+        columns=constants.PRIOR_LAKE_DATA_COLUMNS)
+
+    assert len(items) == 5389
+    for key, val in constants.TEST_PLAKE_ITEM_DICT.items():
+        assert val == items[4596][key]
 
 
 def test_read_shapefile_obscured():
     """
     Tests reading attributes from the shapefile with real values obscured
     """
-    items = swot_reach_node_shp.read_shapefile(
-        constants.TEST_SHAPEFILE_PATH,
+    items = swot_shp.read_shapefile(
+        constants.TEST_REACH_SHAPEFILE_PATH,
         obscure_data=True,
         columns=constants.REACH_DATA_COLUMNS)
 
     assert len(items) == 687
-    for key, val in constants.TEST_ITEM_DICT.items():
+    for key, val in constants.TEST_REACH_ITEM_DICT.items():
         if key == constants.FIELDNAME_WSE:
             assert val != items[2][key]
+
+
+def test_read_benchmarking_data():
+    """
+    Tests reading the benchmarking data
+    """
+    items = swot_shp.load_benchmarking_data()
+
+    assert len(items) == 1199